feat(desktop): M2 页面领域 controller + Workbench 渲染链稳定性修复
领域 controller:
- 新增 CodeController(项目/标签/审批共享状态)与 MarketController(来源/目录/详情)
- CodePage/MarketPage 改为从 controller 读数据,消除重复请求
- MarketController 用 i18n.global.t(useI18n 不能在模块顶层调用)

Workbench 渲染链稳定性:
- 移除 App.vue 主窗口 RouterView 的 out-in Transition 与 WorkbenchShell 内部
  Transition,避免嵌套过渡导致 removed-node patch(parentNode/subTree null)
- setContainerSize 幂等 + ResizeObserver 尺寸取整,阻断反馈循环
- WorkbenchCardHost 改为 provide/inject 传卡片上下文,消除 extraneous props 警告
- 更新 windowLifecycle 契约测试:主窗口不再有 out-in

验证:desktop 测试 33 文件 299 用例全绿,vue-tsc + vite build 通过,
preview 下 Home/Code/Market 路由切换正常渲染。
