+++ apps/desktop/src/pages/CodePage.vue
@@ -9,9 +9,8 @@ import {
   Search,
   X,
 } from '@lucide/vue'
-import { computed, onMounted, ref, watch } from 'vue'
+import { computed, onMounted, ref } from 'vue'
 import { useRouter } from 'vue-router'
-import { api, type ApprovalDto, type ProjectDto } from '@/api'
 import AppHeader from '@/components/AppHeader.vue'
 import FileTreePanel from '@/components/code/FileTreePanel.vue'
 import SearchPanel from '@/components/code/SearchPanel.vue'
@@ -19,165 +18,21 @@ import CodeViewer from '@/components/code/CodeViewer.vue'
 import CodeEditor from '@/components/code/CodeEditor.vue'
 import PatchPreview from '@/components/code/PatchPreview.vue'
 import { UiButton, UiSelect } from '@/components/ui'
-import { useNotifications } from '@/composables/useNotifications'
-
-interface OpenTab {
-  path: string
-  mode: 'view' | 'edit'
-  content?: string
-}
+import { codeController } from '@/controllers/CodeController'
 
 const router = useRouter()
-const { notify, banner, dismissByKey } = useNotifications()
-
-const projects = ref<ProjectDto[]>([])
-const selectedProjectId = ref<string | null>(null)
-const openTabs = ref<OpenTab[]>([])
-const activeTabPath = ref<string | null>(null)
-const approvals = ref<ApprovalDto[]>([])
-const selectedSessionId = ref<string | null>(null)
-const showSearchPanel = ref(true)
-const showPatchPanel = ref(false)
-const patchOriginal = ref('')
-const patchModified = ref('')
-const patchFilePath = ref('')
-const busy = ref(false)
-
-const currentProject = computed(() =>
-  projects.value.find((p) => p.id === selectedProjectId.value) ?? null,
-)
-const currentProjectPath = computed(() => currentProject.value?.path ?? '')
-const activeTab = computed(() =>
-  openTabs.value.find((t) => t.path === activeTabPath.value) ?? null,
-)
-
-async function loadProjects(): Promise<void> {
-  busy.value = true
-  try {
-    projects.value = await api.listProjects()
-    if (!selectedProjectId.value && projects.value.length > 0) {
-      selectedProjectId.value = projects.value[0].id
-    }
-    await loadSession()
-    dismissByKey('code-load')
-  } catch (err) {
-    banner.error({
-      key: 'code-load',
-      title: 'Failed to load projects',
-      message: 'The project list is currently unavailable.',
-      details: err instanceof Error ? err.message : 'Failed to load projects',
-      action: { label: 'Retry', run: () => loadProjects() },
-    })
-  } finally {
-    busy.value = false
-  }
-}
-
-async function loadSession(): Promise<void> {
-  if (!selectedSessionId.value && currentProject.value) {
-    try {
-      const sessions = await api.listSessions(currentProject.value.id)
-      selectedSessionId.value = sessions[0]?.id ?? null
-    } catch {
-      selectedSessionId.value = null
-    }
-  }
-  await loadApprovals()
-}
-
-async function loadApprovals(): Promise<void> {
-  try {
-    const list = await api.listApprovals(selectedSessionId.value ?? undefined)
-    approvals.value = list
-  } catch {
-    approvals.value = []
-  }
-}
-
-function handleFileSelect(path: string): void {
-  const existing = openTabs.value.find((t) => t.path === path)
-  if (existing) {
-    activeTabPath.value = path
-    return
-  }
-  openTabs.value = [...openTabs.value, { path, mode: 'view' }]
-  activeTabPath.value = path
-}
-
-function handleEditFile(path: string, content: string): void {
-  const idx = openTabs.value.findIndex((t) => t.path === path)
-  if (idx !== -1) {
-    openTabs.value[idx] = { ...openTabs.value[idx], mode: 'edit', content }
-  }
-}
-
-function handleCloseTab(path: string): void {
-  const idx = openTabs.value.findIndex((t) => t.path === path)
-  if (idx === -1) return
-  openTabs.value = openTabs.value.filter((t) => t.path !== path)
-  if (activeTabPath.value === path) {
-    activeTabPath.value = openTabs.value[idx]?.path ?? openTabs.value[idx - 1]?.path ?? null
-  }
-}
-
-function handleSwitchTab(path: string): void {
-  activeTabPath.value = path
-}
-
-function handleSwitchToView(path: string): void {
-  const idx = openTabs.value.findIndex((t) => t.path === path)
-  if (idx !== -1) {
-    openTabs.value[idx] = { ...openTabs.value[idx], mode: 'view' }
-  }
-}
-
-function handleApprovalCreated(approval: ApprovalDto): void {
-  approvals.value = [approval, ...approvals.value]
-}
-
-async function decideApproval(approval: ApprovalDto, decision: 'approved' | 'rejected'): Promise<void> {
-  try {
-    await api.decideApproval(approval.id, decision)
-    await loadApprovals()
-  } catch (err) {
-    notify.error(err, { title: 'Failed to decide approval' })
-  }
-}
-
-function handleShowPatch(filePath: string, original: string, modified: string): void {
-  patchFilePath.value = filePath
-  patchOriginal.value = original
-  patchModified.value = modified
-  showPatchPanel.value = true
-}
-
-function handleNewFile(): void {
-  notify.info({
-    title: 'New file',
-    message: 'New file creation requires an approval flow. Use the chat panel to request file creation.',
-  })
-}
-
-function handleRefresh(): void {
-  void loadApprovals()
-}
-
-watch(selectedProjectId, () => {
-  openTabs.value = []
-  activeTabPath.value = null
-  void loadSession()
-})
+const c = codeController
+const activeTab = computed(() => c.activeTab.value)
 
 onMounted(() => {
-  void loadProjects()
+  c.start()
 })
 </script>
 
 <template>
 <main class="shell">
-<!-- Full-width draggable bar for window dragging -->
 <div class="top-drag-bar" />
-<AppHeader :busy="busy" />
+<AppHeader :busy="c.busy.value" />
 
     <section class="code-workspace">
       <!-- Top toolbar -->
@@ -188,14 +43,14 @@ onMounted(() => {
 
         <div class="code-toolbar-project">
           <UiSelect
-            :model-value="selectedProjectId ?? ''"
+            :model-value="c.selectedProjectId.value ?? ''"
             placeholder="Select project..."
             class="h-8 w-64"
-            @update:model-value="selectedProjectId = $event"
+            @update:model-value="c.setProject($event as string)"
           >
             <template #default="{ select, selectedValue }">
               <button
-                v-for="project in projects"
+                v-for="project in c.projects.value"
                 :key="project.id"
                 class="flex w-full items-center gap-2 px-3 py-2 text-left text-sm hover:bg-accent"
                 :class="{ 'bg-accent': selectedValue === project.id }"
@@ -209,19 +64,19 @@ onMounted(() => {
         </div>
 
         <div class="code-toolbar-actions">
-          <UiButton variant="ghost" size="sm" class="h-8" title="New file" @click="handleNewFile">
+          <UiButton variant="ghost" size="sm" class="h-8" title="New file" @click="c.handleNewFile()">
             <FilePlus :size="14" />
             <span>New</span>
           </UiButton>
-          <UiButton variant="ghost" size="sm" class="h-8" title="Refresh" @click="handleRefresh">
+          <UiButton variant="ghost" size="sm" class="h-8" title="Refresh" @click="c.handleRefresh()">
             <RefreshCw :size="14" />
           </UiButton>
           <UiButton
             variant="ghost"
             size="icon"
             class="h-8 w-8"
-            :title="showSearchPanel ? 'Hide search' : 'Show search'"
-            @click="showSearchPanel = !showSearchPanel"
+            :title="c.showSearchPanel.value ? 'Hide search' : 'Show search'"
+            @click="c.showSearchPanel.value = !c.showSearchPanel.value"
           >
             <Search :size="15" />
           </UiButton>
@@ -229,8 +84,8 @@ onMounted(() => {
             variant="ghost"
             size="icon"
             class="h-8 w-8"
-            :title="showPatchPanel ? 'Hide patch' : 'Show patch'"
-            @click="showPatchPanel = !showPatchPanel"
+            :title="c.showPatchPanel.value ? 'Hide patch' : 'Show patch'"
+            @click="c.showPatchPanel.value = !c.showPatchPanel.value"
           >
             <GitCompare :size="15" />
           </UiButton>
@@ -240,20 +95,20 @@ onMounted(() => {
       <!-- Main content area -->
       <div class="code-content">
         <!-- Left panel: file tree + search -->
-        <aside v-if="showSearchPanel" class="code-left-panel">
+        <aside v-if="c.showSearchPanel.value" class="code-left-panel">
           <div class="code-left-section">
             <FileTreePanel
-              :cwd="currentProjectPath"
-              :approvals="approvals"
-              :selected-session-id="selectedSessionId"
-              @select="handleFileSelect"
-              @approval-created="handleApprovalCreated"
+              :cwd="c.currentProjectPath.value"
+              :approvals="c.approvals.value"
+              :selected-session-id="c.selectedSessionId.value"
+              @select="c.handleFileSelect($event)"
+              @approval-created="c.handleApprovalCreated($event)"
             />
           </div>
           <div class="code-left-section code-left-search">
             <SearchPanel
-              :cwd="currentProjectPath"
-              @select="handleFileSelect"
+              :cwd="c.currentProjectPath.value"
+              @select="c.handleFileSelect($event)"
             />
           </div>
         </aside>
@@ -261,17 +116,17 @@ onMounted(() => {
         <!-- Center panel: editor tabs -->
         <main class="code-center-panel">
           <!-- Tab bar -->
-          <div v-if="openTabs.length > 0" class="code-tab-bar">
+          <div v-if="c.openTabs.value.length > 0" class="code-tab-bar">
             <button
-              v-for="tab in openTabs"
+              v-for="tab in c.openTabs.value"
               :key="tab.path"
               class="code-tab"
-              :class="{ active: activeTabPath === tab.path }"
-              @click="handleSwitchTab(tab.path)"
+              :class="{ active: c.activeTabPath.value === tab.path }"
+              @click="c.handleSwitchTab(tab.path)"
             >
               <span class="code-tab-name">{{ tab.path.split(/[\\/]/).pop() }}</span>
               <span class="code-tab-mode">{{ tab.mode }}</span>
-              <span class="code-tab-close" @click.stop="handleCloseTab(tab.path)">
+              <span class="code-tab-close" @click.stop="c.handleCloseTab(tab.path)">
                 <X :size="11" />
               </span>
             </button>
@@ -286,38 +141,38 @@ onMounted(() => {
               <CodeViewer
                 v-if="activeTab.mode === 'view'"
                 :key="`viewer-${activeTab.path}`"
-                :cwd="currentProjectPath"
+                :cwd="c.currentProjectPath.value"
                 :file-path="activeTab.path"
-                @edit="handleEditFile"
+                @edit="c.handleEditFile(activeTab.path, $event)"
               />
               <CodeEditor
                 v-else
                 :key="`editor-${activeTab.path}`"
-                :cwd="currentProjectPath"
+                :cwd="c.currentProjectPath.value"
                 :file-path="activeTab.path"
                 :initial-content="activeTab.content"
-                :selected-session-id="selectedSessionId"
-                :approvals="approvals"
-                @approval-requested="handleApprovalCreated"
-                @saved="handleSwitchToView"
-                @cancel="() => { if (activeTab) handleSwitchToView(activeTab.path) }"
+                :selected-session-id="c.selectedSessionId.value"
+                :approvals="c.approvals.value"
+                @approval-requested="c.handleApprovalCreated($event)"
+                @saved="c.handleSwitchToView(activeTab.path)"
+                @cancel="c.handleSwitchToView(activeTab.path)"
               />
             </template>
           </div>
         </main>
 
         <!-- Right panel: patch preview (optional) -->
-        <aside v-if="showPatchPanel" class="code-right-panel">
+        <aside v-if="c.showPatchPanel.value" class="code-right-panel">
           <PatchPreview
-            v-if="patchFilePath"
-            :cwd="currentProjectPath"
-            :file-path="patchFilePath"
-            :original-content="patchOriginal"
-            :modified-content="patchModified"
-            :selected-session-id="selectedSessionId"
-            :approvals="approvals"
-            @approval-requested="handleApprovalCreated"
-            @cancel="showPatchPanel = false"
+            v-if="c.patchFilePath.value"
+            :cwd="c.currentProjectPath.value"
+            :file-path="c.patchFilePath.value"
+            :original-content="c.patchOriginal.value"
+            :modified-content="c.patchModified.value"
+            :selected-session-id="c.selectedSessionId.value"
+            :approvals="c.approvals.value"
+            @approval-requested="c.handleApprovalCreated($event)"
+            @cancel="c.showPatchPanel.value = false"
           />
           <div v-else class="code-right-empty">
             <GitCompare :size="24" class="text-muted-foreground" />
