+++ apps/desktop/src/workbench/cards/home/TerminalCard.vue
@@ -1,11 +1,14 @@
 <script setup lang="ts">
-import { computed } from 'vue'
+import { computed, inject } from 'vue'
 import TerminalPanel from '@/components/TerminalPanel.vue'
 import { homeController } from '@/controllers/HomeController'
 
 const c = homeController
 
-const visible = computed(() => c.busy.value)
+// Card context is provided by WorkbenchCardHost.
+const active = inject<boolean>('wb:active', false)
+
+const visible = computed(() => active ?? false)
 </script>
 
 <template>
